+++ tests/EndpointTest.php
@@ -12,12 +12,15 @@ class EndpointTest extends TestCase
     {
         $endpoint = new Endpoint('https://laravel.com', 'docs');
 
+        $this->assertSame('https', $endpoint->getScheme());
+        $this->assertSame('laravel.com', $endpoint->getHost());
         $this->assertSame('https://laravel.com', $endpoint->getUri());
         $this->assertSame(['docs'], $endpoint->getPath());
         $this->assertSame([], $endpoint->getQuery());
 
         $this->assertInstanceOf('GuzzleHttp\Psr7\Uri', $endpoint->get());
         $this->assertSame('https://laravel.com/docs', (string) $endpoint->get());
+        $this->assertSame('https://laravel.com/docs', (string) $endpoint);
     }
 
     /** @test */
@@ -25,12 +28,15 @@ class EndpointTest extends TestCase
     {
         $endpoint = new Endpoint('https://laravel.com/', 'docs');
 
+        $this->assertSame('https', $endpoint->getScheme());
+        $this->assertSame('laravel.com', $endpoint->getHost());
         $this->assertSame('https://laravel.com', $endpoint->getUri());
         $this->assertSame(['docs'], $endpoint->getPath());
         $this->assertSame([], $endpoint->getQuery());
 
         $this->assertInstanceOf('GuzzleHttp\Psr7\Uri', $endpoint->get());
         $this->assertSame('https://laravel.com/docs', (string) $endpoint->get());
+        $this->assertSame('https://laravel.com/docs', (string) $endpoint);
     }
 
     /** @test */
@@ -38,12 +44,15 @@ class EndpointTest extends TestCase
     {
         $endpoint = new Endpoint('https://laravel.com', ['docs', '5.4']);
 
+        $this->assertSame('https', $endpoint->getScheme());
+        $this->assertSame('laravel.com', $endpoint->getHost());
         $this->assertSame('https://laravel.com', $endpoint->getUri());
         $this->assertSame(['docs', '5.4'], $endpoint->getPath());
         $this->assertSame([], $endpoint->getQuery());
 
         $this->assertInstanceOf('GuzzleHttp\Psr7\Uri', $endpoint->get());
         $this->assertSame('https://laravel.com/docs/5.4', (string) $endpoint->get());
+        $this->assertSame('https://laravel.com/docs/5.4', (string) $endpoint);
     }
 
     /** @test */
@@ -51,12 +60,15 @@ class EndpointTest extends TestCase
     {
         $endpoint = new Endpoint('https://laravel.com', 'docs', ['search' => 'controller']);
 
+        $this->assertSame('https', $endpoint->getScheme());
+        $this->assertSame('laravel.com', $endpoint->getHost());
         $this->assertSame('https://laravel.com', $endpoint->getUri());
         $this->assertSame(['docs'], $endpoint->getPath());
         $this->assertSame(['search' => 'controller'], $endpoint->getQuery());
 
         $this->assertInstanceOf('GuzzleHttp\Psr7\Uri', $endpoint->get());
         $this->assertSame('https://laravel.com/docs?search=controller', (string) $endpoint->get());
+        $this->assertSame('https://laravel.com/docs?search=controller', (string) $endpoint);
     }
 
     /** @test */
@@ -65,12 +77,15 @@ class EndpointTest extends TestCase
         $endpoint = (new Endpoint('https://laravel.com', 'docs'))
                         ->addQuery(['search' => 'controller', 'page' => 3]);
 
+        $this->assertSame('https', $endpoint->getScheme());
+        $this->assertSame('laravel.com', $endpoint->getHost());
         $this->assertSame('https://laravel.com', $endpoint->getUri());
         $this->assertSame(['docs'], $endpoint->getPath());
         $this->assertSame(['search' => 'controller', 'page' => 3], $endpoint->getQuery());
 
         $this->assertInstanceOf('GuzzleHttp\Psr7\Uri', $endpoint->get());
         $this->assertSame('https://laravel.com/docs?search=controller&page=3', (string) $endpoint->get());
+        $this->assertSame('https://laravel.com/docs?search=controller&page=3', (string) $endpoint);
     }
 
     /** @test */
@@ -78,11 +93,24 @@ class EndpointTest extends TestCase
     {
         $endpoint = new Endpoint(null, 'docs');
 
+        $this->assertSame('', $endpoint->getScheme());
+        $this->assertSame('', $endpoint->getHost());
         $this->assertSame(null, $endpoint->getUri());
         $this->assertSame(['docs'], $endpoint->getPath());
         $this->assertSame([], $endpoint->getQuery());
 
         $this->assertInstanceOf('GuzzleHttp\Psr7\Uri', $endpoint->get());
         $this->assertSame('/docs', (string) $endpoint->get());
+        $this->assertSame('/docs', (string) $endpoint);
+    }
+
+    /**
+     * @test
+     * @expectedException \BadMethodCallException
+     * @expectedExceptionMessage Method [getFullUrl] doesn't exists.
+     */
+    function it_cant_call_unknown_method_should_throw_exception()
+    {
+        (new Endpoint(null, 'docs'))->getFullUrl();
     }
 }